requirements.txt
python-telegram-bot==20.7
firebase-admin==6.5.0
flask==2.3.3